Rename to Authentication::Authenticator
