+++ app/controllers/omniauth_callbacks_controller.rb
@@ -37,11 +37,11 @@ class OmniauthCallbacksController < Devise::OmniauthCallbacksController
 
   def callback_for(provider)
     cta_variant = request.env["omniauth.params"]["state"].to_s
-    @user = AuthorizationService.new(
+    @user = Authentication::Authenticator.call(
       request.env["omniauth.auth"],
       current_user: current_user,
       cta_variant: cta_variant,
-    ).get_user
+    )
 
     if persisted_and_valid?
       # delete legacy session based cookie once the user has logged in again
